Move the code to course_view.py
