+++ app/course_views.py
@@ -14,9 +14,13 @@ from app.utils import (
 from datetime import datetime, timedelta
 from django.db import transaction
 
+from boottest import local_dict
+# TODO：新的本地设置读取都应该使用constants.py中的api引入，有待统一修改
+
 __all__ = [
     'editCourseActivity', 
     'addSingleCourseActivity',
+    'showCourseActivity'
 ]
 
 
@@ -133,3 +137,59 @@ def addSingleCourseActivity(request):
     edit = False  # 前端据此区分是编辑还是创建
 
     return render(request, "lesson_add.html", locals())
+
+
+@login_required(redirect_field_name='origin')
+@utils.check_user_access(redirect_url="/logout/")
+@log.except_captured(source='org_views[showCourseActivity]', record_user=True)
+def showCourseActivity(request):
+    """
+    筛选本学期已结束的课程活动、未开始的课程活动，在课程活动聚合页面进行显示。
+    """
+
+    # Sanity check and start a html_display.
+    user = request.user
+    valid, user_type, html_display = utils.check_user_type(request.user)
+    me = get_person_or_org(user, user_type)  # 获取自身
+
+    
+    if user_type == "Person":
+        return redirect(message_url(wrong('只有书院课程组织才能查看此页面!')))
+
+    type_name = me.otype.otype_name
+    # TODO：“书院课程”从json读，constants.py可能应该为此增加一项
+    if type_name != "书院课程":
+        return redirect(message_url(wrong('只有书院课程组织才能查看此页面!')))
+
+    all_activity_list = (
+        Activity.objects
+        .filter(organization_id=me)
+        .filter(category=Activity.ActivityCategory.COURSE)
+        .order_by("-start")
+    )
+
+    future_activity_list = (
+        all_activity_list.filter(
+            status__in=[
+                Activity.Status.REVIEWING,
+                Activity.Status.APPLYING,
+                Activity.Status.WAITING,
+                Activity.Status.PROGRESSING,
+            ]
+        )
+    )
+
+    finished_activity_list = (
+        all_activity_list
+        .filter(status=Activity.Status.END)
+        .filter(
+            year=int(local_dict["semester_data"]["year"]),
+            semester__contains=local_dict["semester_data"]["semester"]
+        )
+        .order_by("-end")
+    ) # 本学期的已结束活动 
+
+    bar_display = utils.get_sidebar_and_navbar(request.user, navbar_name="我的活动")
+
+    return render(request, "org_show_course_activity.html", locals())
+
